[device-bootloaders] Adding Power Event status trace support

+++ device-bootloaders/src/ServiceProcessorBL2/services/perf_mgmt.c
@@ -197,9 +197,9 @@ int32_t Get_Minion_Frequency(void)
 *      int                    Return status
 *
 ***********************************************************************/
-void Update_Minion_Frequency_Global_Reg(uint32_t new_freq)
+void Update_Minion_Frequency_Global_Reg(int32_t new_freq)
 {
-    get_soc_perf_reg()->asic_frequency.minion_shire_mhz = new_freq;
+    get_soc_perf_reg()->asic_frequency.minion_shire_mhz = (uint32_t)new_freq;
 }
 
 /************************************************************************
